# swagger.yaml
openapi: 3.1.0
info:
  title: Mail System API
  description: "REST API for the THM Web Technologies mail system. The API supports\
    \ JWT authentication, user management, drafting and sending mails, incoming mail\
    \ retrieval, and protected attachment downloads."
  contact:
    name: Mail System Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local backend server
- url: http://localhost
  description: Docker Compose frontend reverse proxy
...
paths:
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get a user
      description: Returns the user profile for the supplied id.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: "User returned successfully. If the id is unknown, the current\
            \ implementation returns an empty response body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
    put:
      tags:
      - Users
      summary: Update the authenticated user
      description: Updates the profile of the authenticated user. Users can only update
        their own profile.
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
        required: true
      responses:
        "200":
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: "The request body, multipart payload, or query parameters are\
            \ invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "404":
          description: The requested resource does not exist or is not visible to
            the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "409":
          description: The request conflicts with an existing resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Users
      summary: Delete the authenticated user
      description: Deletes the authenticated user's own account.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: User deleted successfully.
        "404":
          description: The requested resource does not exist or is not visible to
            the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
...
  /api/users/ensure:
    post:
      tags:
      - Users
      summary: Ensure an external contact
      description: "Looks up a user by normalized email. If none exists, creates an\
        \ external contact user with generated credentials."
      operationId: ensureExternalUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnsureUserRequest"
        required: true
      responses:
        "200":
          description: Existing or newly created contact returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnsureUserResponse"
        "400":
          description: "The request body, multipart payload, or query parameters are\
            \ invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
...
  /api/users:
    get:
      tags:
      - Users
      summary: List users
      description: Returns all users known to the application. This list is used to
        select recipients while composing mails.
      operationId: listUsers
      responses:
        "200":
          description: Users returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
...
  /api/attachments/{attachmentId}:
    get:
      tags:
      - Attachments
      summary: Download an attachment
      description: Streams an attachment if the authenticated user is the sender or
        recipient of the related mail. Unauthorized access is hidden as 404.
      operationId: downloadAttachment
      parameters:
      - name: attachmentId
        in: path
        description: Attachment id.
        required: true
        schema:
          type: string
          format: uuid
        example: 3f75cd3b-3caa-4765-bdc3-c40279a1975a
      responses:
        "200":
          description: Attachment file stream returned successfully. The response
            content type is based on the stored attachment metadata.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        "404":
          description: The requested resource does not exist or is not visible to
            the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "401":
          description: "Unauthorized. The bearer token is missing, invalid, or expired."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
      security:
      - bearerAuth: []
components:
  schemas:
    UserUpdate:
      type: object
      description: Request payload for updating the authenticated user's profile.
      properties:
        firstName:
          type: string
          description: Updated first name.
          example: Ameline
          maxLength: 2147483647
          minLength: 1
        lastName:
          type: string
          description: Updated last name.
          example: Allanson
          maxLength: 2147483647
          minLength: 1
        email:
          type: string
          format: email
          description: Updated email address. Must not already be used by another
            user.
          example: ameline.allanson@example.com
      required:
      - email
      - firstName
      - lastName
    AppError:
      type: object
      description: Standard error response returned by the backend.
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code.
          example: 404
        message:
          type:
          - string
          - "null"
          description: Human-readable error message.
          example: Mail not found
      required:
      - status
    UserDTO:
      type: object
      description: Public user profile returned by the API.
      properties:
        id:
          type:
          - string
          - "null"
          format: uuid
          description: Unique user id.
          example: 6f1f9368-d279-4a6e-993f-f0618767eeb8
        firstName:
          type: string
          description: User's first name.
          example: Ameline
        lastName:
          type: string
          description: User's last name.
          example: Allanson
        email:
          type: string
          description: User's email address.
          example: aallanson@example.com
      required:
      - email
      - firstName
      - lastName
...
    EnsureUserRequest:
      type: object
      description: Request payload for finding or creating an external contact user.
      properties:
        email:
          type: string
          format: email
          description: Email address of the external contact. The server normalizes
            it before lookup.
          example: customer@example.com
          minLength: 1
        firstName:
          type:
          - string
          - "null"
          description: "Optional first name. If omitted or blank, the server uses\
            \ `External`."
          example: Chris
        lastName:
          type:
          - string
          - "null"
          description: "Optional last name. If omitted or blank, the server uses `User`."
          example: Customer
      required:
      - email
    EnsureUserResponse:
      type: object
      description: External contact user returned after lookup or creation.
      properties:
        id:
          type:
          - string
          - "null"
          format: uuid
          description: Unique contact user id.
          example: 4ff9c0df-c0f5-4fa2-a7d5-6e517f22186c
        firstName:
          type: string
          description: Contact first name.
          example: External
        lastName:
          type: string
          description: Contact last name.
          example: User
        email:
          type: string
          description: Contact email address.
          example: customer@example.com
      required:
      - email
      - firstName
      - lastName
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "JWT access token returned by `POST /api/login` or `POST /api/register`.\
        \ Use the `Authorization: Bearer <token>` header."
      scheme: bearer
      bearerFormat: JWT